execution/protocol/rules/ethash: drop in-memory + on-disk cache
Post-Merge ethash is only used for tests and historical re-execution, so
the LRU cache + mmap-backed disk persistence around the verification
cache and mining DAG no longer earns its keep. verifySeal now allocates
a fresh cache per call and runs hashimotoLight directly. The dataset
(fulldag) verification path is also gone.

Removes the related ethashcfg fields (CachesInMem, CachesLockMmap,
DatasetDir, DatasetsInMem, DatasetsOnDisk, DatasetsLockMmap), their
defaults, CLI flags, and the OS-specific DAG-dir init.

ethash.go: 594 -> 171 lines.
