+++ execution/protocol/rules/ethash/rules.go
@@ -24,7 +24,6 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
-	"runtime"
 	"time"
 
 	mapset "github.com/deckarep/golang-set/v2"
@@ -336,57 +335,37 @@ func CalcDifficulty(config *chain.Config, time, parentTime uint64, parentDifficu
 // VerifySeal implements rules.Engine, checking whether the given block satisfies
 // the PoW difficulty requirements.
 func (ethash *Ethash) VerifySeal(_ rules.ChainHeaderReader, header *types.Header) error {
-	return ethash.verifySeal(header, false)
+	return ethash.verifySeal(header)
 }
 
-// verifySeal checks whether a block satisfies the PoW difficulty requirements,
-// either using the usual ethash cache for it, or alternatively using a full DAG
-// to make remote mining fast.
-func (ethash *Ethash) verifySeal(header *types.Header, fulldag bool) error { //nolint:unparam
+// verifySeal checks whether a block satisfies the PoW difficulty requirements
+// using a freshly generated ethash verification cache. The cache is not retained:
+// post-Merge ethash is only used for tests and historical re-execution, so the
+// memory savings outweigh recomputation cost.
+func (ethash *Ethash) verifySeal(header *types.Header) error {
 	// If we're running a shared PoW, delegate verification to it
 	if ethash.shared != nil {
-		return ethash.shared.verifySeal(header, fulldag)
+		return ethash.shared.verifySeal(header)
 	}
 
 	// Ensure that we have a valid difficulty for the block
 	if header.Difficulty.Sign() <= 0 {
 		return errInvalidDifficulty
 	}
-	// Recompute the digest and PoW values
 	number := header.Number.Uint64()
+	epoch := number / epochLength
 
-	var (
-		digest []byte
-		result []byte
-	)
-	// If fast-but-heavy PoW verification was requested, use an ethash dataset
-	if fulldag {
-		dataset := ethash.dataset(number, true)
-		if dataset.generated() {
-			digest, result = hashimotoFull(dataset.dataset, ethash.SealHash(header).Bytes(), header.Nonce.Uint64())
-
-			// Datasets are unmapped in a finalizer. Ensure that the dataset stays alive
-			// until after the call to hashimotoFull so it's not unmapped while being used.
-			runtime.KeepAlive(dataset)
-		} else {
-			// Dataset not yet generated, don't hang, use a cache instead
-			fulldag = false
-		}
+	csize := cacheSize(number)
+	dsize := datasetSize(number)
+	if ethash.config.PowMode == ethashcfg.ModeTest {
+		csize = 1024
+		dsize = 32 * 1024
 	}
-	// If slow-but-light PoW verification was requested (or DAG not yet ready), use an ethash cache
-	if !fulldag {
-		cache := ethash.cache(number)
+	cache := make([]uint32, csize/4)
+	generateCache(cache, epoch, seedHash(number))
 
-		size := datasetSize(number)
-		if ethash.config.PowMode == ethashcfg.ModeTest {
-			size = 32 * 1024
-		}
-		digest, result = hashimotoLight(size, cache.cache, ethash.SealHash(header).Bytes(), header.Nonce.Uint64())
+	digest, result := hashimotoLight(dsize, cache, ethash.SealHash(header).Bytes(), header.Nonce.Uint64())
 
-		// Caches are unmapped in a finalizer. Ensure that the cache stays alive
-		// until after the call to hashimotoLight so it's not unmapped while being used.
-		runtime.KeepAlive(cache)
-	}
 	// Verify the calculated values against the ones provided in the header
 	if !bytes.Equal(header.MixDigest[:], digest) {
 		return errInvalidMixDigest
